ignore NS PSa labels even for live resource lookups
The Pod Security admission labels on a namespace would interfere
with the admission that's run on the workloads, ignore it to be
able to get real results per given workload.

Example: An NS has privileged enforcement set. The admission runs
on a pod in the NS, but all admission levels automatically allow
since the NS already is privileged.